requirements1.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.5
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.14
django-birthday==0.1.4
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-jazzmin==2.5.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
asgiref==3.5.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.5
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-birthday==0.1.4
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-jazzmin==2.5.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
asgiref==3.5.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
cffi==1.15.0
channels==3.0.5
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-birthday==0.1.4
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-jazzmin==2.5.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
PyYAML==6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
